assign espaloma charges to single atom molecules
